Correct table names in inst_data.xml
Some data was ending up going to the wrong table because the tables
names in inst_data.xml did not correspond to the file data/name. This is
probably caused by a bad copy/paste and forgetting to change the table
name in the xml. The commit fixes that error.